# Optimize submission of tools that output collections.