@jmchilton 2020-06-03
Sounds good - let me know how it goes? Did that disabling that legacy setting help things at all?