@bgruening 2020-06-03
```
Jun 03 19:44:58 sn04.bi.uni-freiburg.de python[3074033]: galaxy.workflow.run DEBUG 2020-06-03 19:44:58,142 Marking step 758717 outputs of invocation 99366 delayed (dependent step [758716] delayed, so this step must be delayed)
Jun 03 19:44:58 sn04.bi.uni-freiburg.de python[3074033]: galaxy.workflow.run DEBUG 2020-06-03 19:44:58,144 Marking step 758718 outputs of invocation 99366 delayed (dependent step [758717] delayed, so this step must be delayed)
Jun 03 19:44:58 sn04.bi.uni-freiburg.de python[3074033]: galaxy.workflow.scheduling_manager DEBUG 2020-06-03 19:44:58,157 Workflow invocation [99366] scheduled
Jun 03 19:44:58 sn04.bi.uni-freiburg.de python[3074033]: galaxy.workflow.scheduling_manager DEBUG 2020-06-03 19:44:58,157 Attempting to schedule workflow invocation [(99367,)]
Jun 03 19:44:58 sn04.bi.uni-freiburg.de python[3074033]: galaxy.workflow.scheduling_manager DEBUG 2020-06-03 19:44:58,185 Attempting to schedule workflow invocation [(99429,)]
Jun 03 19:44:58 sn04.bi.uni-freiburg.de python[3073995]: galaxy.tools.actions INFO 2020-06-03 19:44:58,889 Handled collection output named output_paired_coll for tool toolshed.g2.bx.psu.edu/repos/iuc/fastp/fastp/0.20.1+galaxy0 (4700.853 ms)
Jun 03 19:44:58 sn04.bi.uni-freiburg.de python[3073995]: galaxy.tools.actions INFO 2020-06-03 19:44:58,952 Handled output named report_html for tool toolshed.g2.bx.psu.edu/repos/iuc/fastp/fastp/0.20.1+galaxy0 (62.085 ms)
Jun 03 19:44:58 sn04.bi.uni-freiburg.de python[3073995]: galaxy.tools.actions INFO 2020-06-03 19:44:58,996 Added output datasets to history (43.274 ms)
Jun 03 19:44:59 sn04.bi.uni-freiburg.de python[3073995]: galaxy.tools.actions INFO 2020-06-03 19:44:59,001 Setup for job Job[unflushed,tool_id=toolshed.g2.bx.psu.edu/repos/iuc/fastp/fastp/0.20.1+galaxy0] complete, ready to be enqueued (4.951 ms)
Jun 03 19:44:59 sn04.bi.uni-freiburg.de python[3073995]: galaxy.web_stack.handlers INFO 2020-06-03 19:44:59,086 Flushed transaction for Job[id=8549119,tool_id=toolshed.g2.bx.psu.edu/repos/iuc/fastp/fastp/0.20.1+galaxy0] (84.940 ms)
Jun 03 19:44:59 sn04.bi.uni-freiburg.de python[3073995]: galaxy.web_stack.handlers INFO 2020-06-03 19:44:59,086 (Job[id=8549119,tool_id=toolshed.g2.bx.psu.edu/repos/iuc/fastp/fastp/0.20.1+galaxy0]) Handler '_default_' assigned using 'db-skip-locked' assignment method
Jun 03 19:44:59 sn04.bi.uni-freiburg.de python[3073995]: galaxy.tools.execute DEBUG 2020-06-03 19:44:59,087 Tool [toolshed.g2.bx.psu.edu/repos/iuc/fastp/fastp/0.20.1+galaxy0] created job [8549119] (4955.356 ms)
Jun 03 19:45:00 sn04.bi.uni-freiburg.de python[3074033]: galaxy.workflow.run DEBUG 2020-06-03 19:45:00,399 Marking step 273942 outputs of invocation 99429 delayed (dependent collection [285700] not yet populated with datasets)
Jun 03 19:45:00 sn04.bi.uni-freiburg.de python[3074033]: galaxy.workflow.run DEBUG 2020-06-03 19:45:00,400 Marking step 273943 outputs of invocation 99429 delayed (dependent collection [285700] not yet populated with datasets)
```

Unfortunately, this does not look better, but maybe it is 19.09 that does not like the patch.

> Did that disabling that legacy setting help things at all?

Yes, but I think in different ways. Zergling pressure is less and our uwsgi-queue-full-errors seem to be less, subjectively.